Divide into two program files
